NF: extract exceptions from MigrateUserData
