+++ AnkiDroid/src/test/java/com/ichi2/anki/servicelayer/scopedstorage/DeleteEmptyDirectoryTest.kt
@@ -17,7 +17,7 @@
 package com.ichi2.anki.servicelayer.scopedstorage
 
 import com.ichi2.anki.model.Directory
-import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.MigrateUserData.DirectoryNotEmptyException
+import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.DirectoryNotEmptyException
 import com.ichi2.compat.Test21And26
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.hasSize
